fix:close hamburger icon issue
