+++ src/AppContext.jsx
@@ -7,7 +7,6 @@ const AppProvider = ({ children }) => {
 
   const handleShowContactForm = () => document.getElementById('my_modal_1').showModal()
   const handleHideContactForm = () => document.getElementById('my_modal_1').close()
-
   const handleHideNavMenu = () => document.getElementById('dropdown-nav').close();
 
   return (
