+++ src/components/Navbar/Navbar.jsx
@@ -1,21 +1,17 @@
 import { useState,useContext } from 'react';
+
 import { useNavigate, NavLink, useLocation } from 'react-router-dom';
 
 import AppContext from '../../AppContext';
-import './navbar.css';
-import img from '../../assets/footer.png';
-//import { Icon } from '@iconify/react';
-import NavMobile from '../NavMobile';
+import NavMobile from './NavMobile';
 import { text } from '../../utils/text';
+import img from '../../assets/footer.png';
+import './navbar.css';
+
 
 function Navbar() {
   const navigate = useNavigate();
   const { language, setLanguage, handleShowContactForm } = useContext(AppContext);
- /*  const [isMenuOpen, setIsMenuOpen] = useState(false); */
-
-  /* const toggleMenu = () => {
-    setIsMenuOpen(!isMenuOpen);
-  } */
   const currentPage = useLocation();
   const [activePage, setActivePage] = useState(currentPage.pathname);
   
@@ -89,51 +85,6 @@ function Navbar() {
       <div className='dropdown dropdown-bottom dropdown-end lg:hidden'>
         <NavMobile />
       </div>
-
-      {/* Mobile burger menu */}
-      {/* This does not work on ONLY on safari */}
-      {/* <div className='dropdown dropdown-bottom dropdown-end lg:hidden'>
-        <button onClick={toggleMenu} className='m-1 btn bg-inherit border-none'>
-          {isMenuOpen ? ( */}
-            {/* Change this icon to an 'X' when menu is open */ }
-            {/* <Icon icon='iconamoon:close' width='30' height='30' />
-          ) : ( */}
-            {/* Display the burger icon when menu is closed */}
-            {/* <Icon icon='iconamoon:menu-burger-horizontal-light' width='30' height='30' />
-          )}
-        </button> */}
-
-       {/*  <nav className={`nav-menu ${isMenuOpen ? 'open' : ''}`}>
-          <ul className='p-2 shadow menu dropdown-content z-50 bg-base-100 rounded-box w-52'>
-            <li> */}
-              {/* Mobile Contact button */}
-             {/*  <button
-                className='px-4 py-2 rounded-3xl bg-blue text-lg font-semibold text-white self-center m-4'
-                onClick={() => { handleShowContactForm(); toggleMenu() }}
-              >
-                {text.contactUs[language]}
-              </button>
-            </li> */}
-            {/* Mobile change language button */}
-            {/* <li>
-              <button onClick={() => { handleChangeLanguage(); toggleMenu() }}>{text.changeLanguage[language]}</button>
-            </li> */}
-            {/* Mobile navNavLinks */}
-            {/* <li>
-              <button onClick={() => { navigate('/'); toggleMenu() }}>{text.home[language]}</button>
-            </li>
-            <li>
-              <button onClick={() => { navigate('/about'); toggleMenu() }}>{text.about[language]}</button>
-            </li>
-            <li>
-              <button onClick={() => { navigate('/news'); toggleMenu() }}>{text.news[language]}</button>
-            </li>
-            <li>
-              <button onClick={() => { navigate('/events'); toggleMenu() }}>{text.events[language]}</button>
-            </li>
-          </ul>
-        </nav>
-      </div> */}
     </div>
   )
 }
